decrease pl version

+++ models/Exposure.py
@@ -147,7 +147,7 @@ class Exposure(pl.LightningModule):
         #print(f"prediction_label shape: {prediction_label.shape}")
 
         loss = self.loss_fn(prediction_label, batch['label'])
-        a = self.confusion_matrix(prediction_label, batch['label'])
+        a = self.confusion_matrix(prediction_label, batch['label'].astype(int))
         print(a)
 
         #loss = F.cross_entropy(prediction_labels['neutral'].sigmoid(), batch['neutral'])
